pyproject.toml
[tool.poetry]
name = "event-api"
version = "0.1.0"
description = "Technical test for tusdatos.co"
authors = ["iKenshu <kenshumorales@gmail.com>"]
packages = [
    { include = "event_api" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
SQLAlchemy = "^2.0.36"
elasticsearch = "^8.16.0"
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
uvicorn = "^0.32.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
graphql-core = "^3.2.5"
strawberry-graphql = "^0.253.0"
httpx = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.19"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.4"
coverage = "^7.6.8"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    "migrations",
]
line-length = 80
indent-width = 4
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = [
    "."
]
addopts = "--cov=. --cov-report=term-missing"
